+++ src/aws-api-mcp-server/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Fetch embedding model from AWS instead of Hugging Face (#1127)
+
 ### Fixed
 
 - Use region from profile specified in cli command (#1123)